+++ src/views/Projetos.vue
@@ -1,53 +1,19 @@
 <template>
-    <section class="projetos">
-        <h1 class="title">Projetos</h1>
-        <router-link to="/projetos/novo" class="button">
-            <span class="icon is-small">
-                <i class="fas fa-plus"></i>
-            </span>
-            <span>Novo projeto</span>
-        </router-link>
-        <table class="table is-fullwidth">
-            <thead>
-                <tr>
-                    <th>ID</th>
-                    <th>Nome</th>
-                    <th>Ações</th>
-                </tr>
-            </thead>
-            <tbody>
-                <tr v-for="projeto in projetos" :key="projeto.id">
-                    <td>{{ projeto.id }}</td>
-                    <td>{{ projeto.nome }}</td>
-                    <td>
-                        <router-link :to="`/projetos/${projeto.id}`" class="button">
-                            <span class="icon is-small">
-                                <i class="fas fa-pencil-alt"></i>
-                            </span>
-                        </router-link>
-                    </td>
-                </tr>
-            </tbody>
-        </table>
-    </section>
-    
+    <div class="projetos">
+        <h1 class="title">
+            Projetos
+        </h1>
+        <RouterView />
+    </div>
+
 </template>
 
 <script lang="ts">
-import { useStore } from '@/store';
-import { computed, defineComponent } from 'vue';
-
+import { defineComponent } from 'vue';
 
 export default defineComponent({
-    name: 'Projetos',
-    setup() {
-        const store = useStore()
-        return {
-            projetos: computed(() => store.state.projetos)
-        }
-    }
+    name: "Projetos"
 })
-
 </script>
 
 <style scoped>
