+++ src/views/Projetos/Formulario.vue
@@ -1,7 +1,5 @@
 <template>
-    <section class="projetos">
-        <h1 class="title">Projetos</h1>
-
+    <section>
         <form @submit.prevent="salvar">
             <div class="field">
                 <label for="nomeDoProjeto" class="label">
@@ -69,9 +67,3 @@ export default defineComponent({
 
 </script>
 
-<style scoped>
-.projetos {
-    padding: 1.25rem;
-}
-
-</style>
